requirements.txt
asgiref==3.4.1
Django==3.2.11
django-cors-headers==3.11.0
djangorestframework==3.13.1
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
